Add version namespace to accommodate future PDF versions
