+++ modules/appeals_api/app/services/appeals_api/pdf_construction/notice_of_disagreement/v1/form_data.rb
@@ -2,7 +2,7 @@
 
 module AppealsApi
   module PdfConstruction
-    module NoticeOfDisagreement
+    module NoticeOfDisagreement::V1
       class FormData
         def initialize(notice_of_disagreement)
           @notice_of_disagreement = notice_of_disagreement
