+++ modules/appeals_api/spec/services/appeals_api/pdf_construction/pages/hearing_type_and_additional_issues_spec.rb
@@ -10,7 +10,7 @@ module AppealsApi::PdfConstruction::NoticeOfDisagreement
       it 'returns the same object that it received' do
         pdf = Prawn::Document.new
 
-        resulted_pdf = described_class.new(pdf, FormData.new(notice_of_disagreement)).build!
+        resulted_pdf = described_class.new(pdf, V1::FormData.new(notice_of_disagreement)).build!
 
         expect(resulted_pdf).to eq(pdf)
       end
@@ -18,7 +18,7 @@ module AppealsApi::PdfConstruction::NoticeOfDisagreement
       it 'starts a new pdf page' do
         pdf = Prawn::Document.new
 
-        expect { described_class.new(pdf, FormData.new(notice_of_disagreement)).build! }
+        expect { described_class.new(pdf, V1::FormData.new(notice_of_disagreement)).build! }
           .to change { pdf.page_count }.by(1)
       end
     end
